Bug Fixes, version bump

+++ app/Services/Payment/DeletePayment.php
@@ -11,11 +11,6 @@
 
 namespace App\Services\Payment;
 
-use App\Exceptions\PaymentRefundFailed;
-use App\Factory\CreditFactory;
-use App\Factory\InvoiceItemFactory;
-use App\Models\Activity;
-use App\Models\CompanyGateway;
 use App\Models\Credit;
 use App\Models\Invoice;
 use App\Models\Payment;
@@ -36,11 +31,16 @@ class DeletePayment
 
     public function run()
     {
+        if ($this->payment->is_deleted) {
+            return $this->payment;
+        }
+
         return $this->setStatus(Payment::STATUS_CANCELLED) //sets status of payment
             ->updateCreditables() //return the credits first
             ->adjustInvoices()
             ->updateClient()
             ->deletePaymentables()
+            ->cleanupPayment()
             ->save();
     }
 
@@ -51,6 +51,15 @@ class DeletePayment
     //set refunded to amount
 
     //set applied amount to 0
+
+    private function cleanupPayment()
+    {
+        $this->payment->is_deleted = true;
+        // $entity->save();
+        $this->payment->delete();
+        
+        return $this;
+    }
 
     private function deletePaymentables()
     {
